Refactoring inner classes out to module level to ease testing.
