+++ android/src/main/java/org/mozilla/osmdroid/tileprovider/modules/MapTileModuleProviderBase.java
@@ -1,10 +1,7 @@
 package org.mozilla.osmdroid.tileprovider.modules;
 
-import android.graphics.drawable.Drawable;
-
 import org.mozilla.mozstumbler.service.AppGlobals;
 import org.mozilla.mozstumbler.service.core.logging.Log;
-import org.mozilla.osmdroid.tileprovider.ExpirableBitmapDrawable;
 import org.mozilla.osmdroid.tileprovider.MapTile;
 import org.mozilla.osmdroid.tileprovider.MapTileRequestState;
 import org.mozilla.osmdroid.tileprovider.constants.OSMConstants;
@@ -42,7 +39,7 @@ public abstract class MapTileModuleProviderBase implements OSMConstants {
 
 	/**
 	 * It is expected that the implementation will construct an internal member which internally
-	 * implements a {@link TileLoader}. This method is expected to return a that internal member to
+	 * implements a {@link AbstractTileLoader}. This method is expected to return a that internal member to
 	 * methods of the parent methods.
 	 *
 	 * @return the internal member of this tile provider.
@@ -150,7 +147,7 @@ public abstract class MapTileModuleProviderBase implements OSMConstants {
 		}
 	}
 
-	private void clearQueue() {
+	void clearQueue() {
 		synchronized (mQueueLockObject) {
 			mPending.clear();
 			mWorking.clear();
@@ -176,138 +173,7 @@ public abstract class MapTileModuleProviderBase implements OSMConstants {
 		}
 	}
 
-	/**
-	 * Load the requested tile. An abstract internal class whose objects are used by worker threads
-	 * to acquire tiles from servers. It processes tiles from the 'pending' set to the 'working' set
-	 * as they become available. The key unimplemented method is 'loadTile'.
-	 */
-	protected abstract class TileLoader implements Runnable {
-
-		/**
-		 * Load the requested tile.
-		 *
-		 * @return the tile if it was loaded successfully, or null if failed to
-		 *         load and other tile providers need to be called
-		 * @param pState
-		 * @throws CantContinueException
-		 */
-		protected abstract Drawable loadTile(MapTileRequestState pState)
-				throws CantContinueException;
-
-		protected void onTileLoaderInit() {
-			// Do nothing by default
-		}
-
-		protected void onTileLoaderShutdown() {
-			// Do nothing by default
-		}
-
-		protected MapTileRequestState nextTile() {
-
-			synchronized (mQueueLockObject) {
-				MapTile result = null;
-
-				// get the most recently accessed tile
-				// - the last item in the iterator that's not already being
-				// processed
-				Iterator<MapTile> iterator = mPending.keySet().iterator();
-
-				// TODO this iterates the whole list, make this faster...
-				while (iterator.hasNext()) {
-					final MapTile tile = iterator.next();
-					if (!mWorking.containsKey(tile)) {
-						if (DEBUG_TILE_PROVIDERS) {
-							Log.d(LOG_TAG, "TileLoader.nextTile() on provider: " + getName()
-									+ " found tile in working queue: " + tile);
-						}
-						result = tile;
-					}
-				}
-
-				if (result != null) {
-					if (DEBUG_TILE_PROVIDERS) {
-						Log.d(LOG_TAG, "TileLoader.nextTile() on provider: " + getName()
-								+ " adding tile to working queue: " + result);
-					}
-					mWorking.put(result, mPending.get(result));
-				}
-
-				return (result != null ? mPending.get(result) : null);
-			}
-		}
-
-		/**
-		 * A tile has loaded.
-		 */
-		protected void tileLoaded(final MapTileRequestState pState, final Drawable pDrawable) {
-			if (DEBUG_TILE_PROVIDERS) {
-				Log.d(LOG_TAG, "TileLoader.tileLoaded() on provider: " + getName() + " with tile: "
-						+ pState.getMapTile());
-			}
-			removeTileFromQueues(pState.getMapTile());
-			pState.getCallback().mapTileRequestCompleted(pState, pDrawable);
-		}
-
-		/**
-		 * A tile has loaded but it's expired.
-		 * Return it <b>and</b> send request to next provider.
-		 */
-		protected void tileLoadedExpired(final MapTileRequestState pState, final Drawable pDrawable) {
-			if (DEBUG_TILE_PROVIDERS) {
-				Log.d(LOG_TAG, "TileLoader.tileLoadedExpired() on provider: " + getName()
-						+ " with tile: " + pState.getMapTile());
-			}
-			removeTileFromQueues(pState.getMapTile());
-			pState.getCallback().mapTileRequestExpiredTile(pState, pDrawable);
-		}
-
-		protected void tileLoadedFailed(final MapTileRequestState pState) {
-			if (DEBUG_TILE_PROVIDERS) {
-				Log.d(LOG_TAG, "TileLoader.tileLoadedFailed() on provider: " + getName()
-						+ " with tile: " + pState.getMapTile());
-			}
-			removeTileFromQueues(pState.getMapTile());
-			pState.getCallback().mapTileRequestFailed(pState);
-		}
-
-		/**
-		 * This is a functor class of type Runnable. The run method is the encapsulated function.
-		 */
-		@Override
-		final public void run() {
-
-			onTileLoaderInit();
-
-			MapTileRequestState state;
-			Drawable result = null;
-			while ((state = nextTile()) != null) {
-				if (DEBUG_TILE_PROVIDERS) {
-					Log.d(LOG_TAG, "TileLoader.run() processing next tile: " + state.getMapTile());
-				}
-				try {
-					result = null;
-					result = loadTile(state);
-				} catch (final CantContinueException e) {
-					Log.e(LOG_TAG, "Tile loader can't continue: " + state.getMapTile(), e);
-					clearQueue();
-				} catch (final Throwable e) {
-					Log.e(LOG_TAG, "Error downloading tile: " + state.getMapTile(), e);
-				}
-
-				if (result == null) {
-					tileLoadedFailed(state);
-				} else if (ExpirableBitmapDrawable.isDrawableExpired(result)) {
-					tileLoadedExpired(state, result);
-				} else {
-					tileLoaded(state, result);
-				}
-			}
-
-			onTileLoaderShutdown();
-		}
-	}
-
-	/**
+    /**
 	 * Thrown by a tile provider module in TileLoader.loadTile() to signal that it can no longer
 	 * function properly. This will typically clear the pending queue.
 	 */
